Refactor: Move translation_tool.py to the chatbot module
Moved `plugin/framework/translation_tool.py` to `plugin/modules/chatbot/translation_tool.py` and updated all relevant references in the codebase. Corresponding test files were also moved to the chatbot tests directory to keep the structure coherent.
