+++ plugin/contrib/aihordeclient/__init__.py
@@ -29,7 +29,7 @@ from time import sleep
 from typing import Any, Dict, List, Union
 from urllib.error import HTTPError, URLError
 from urllib.request import Request
-from plugin.framework.translation_tool import opustm_hf_translate, OPUSTM_SOURCE_LANGUAGES  # noqa F401
+from plugin.modules.chatbot.translation_tool import opustm_hf_translate, OPUSTM_SOURCE_LANGUAGES  # noqa F401
 
 from plugin.modules.http.requests import sync_request
 from plugin.modules.http.errors import format_error_message
