+++ plugin/modules/chatbot/translation_tool.py
